Move schema keys to a separate class
