+++ nacc_form_validator/quality_check.py
@@ -6,51 +6,8 @@ from cerberus.errors import DocumentErrorTree
 from cerberus.schema import SchemaError
 
 from nacc_form_validator.datastore import Datastore
-from nacc_form_validator.nacc_validator import (
-    CustomErrorHandler,
-    NACCValidator,
-    ValidationException,
-)
-
-
-class SchemaDefs:
-    """Class to store schema attribute labels."""
-
-    TYPE = "type"
-    OP = "op"
-    IF_OP = "if_op"
-    THEN_OP = "then_op"
-    ELSE_OP = "else_op"
-    IF = "if"
-    THEN = "then"
-    ELSE = "else"
-    META = "meta"
-    ERRMSG = "errmsg"
-    ORDERBY = "orderby"
-    CONSTRAINTS = "constraints"
-    PREV_OP = "prev_op"
-    CURR_OP = "curr_op"
-    CURRENT = "current"
-    PREVIOUS = "previous"
-    CRR_DATE = "current_date"
-    CRR_YEAR = "current_year"
-    CRR_MONTH = "current_month"
-    CRR_DAY = "current_day"
-    PREV_RECORD = "previous_record"
-    FORMULA = "formula"
-    INDEX = "index"
-    FORMATTING = "formatting"
-    COMPARATOR = "comparator"
-    BASE = "base"
-    ADJUST = "adjustment"
-    IGNORE_EMPTY = "ignore_empty"
-    BIRTH_MONTH = 'birth_month'
-    BIRTH_DAY = 'birth_day'
-    BIRTH_YEAR = 'birth_year'
-    COMPARE_TO = "compare_to"
-    SWAP_ORDER = "swap_order"
-    FUNCTION_NAME = 'name'
-    FUNCTION_ARGS = 'args'
+from nacc_form_validator.errors import CustomErrorHandler
+from nacc_form_validator.nacc_validator import NACCValidator, ValidationException
 
 
 class QualityCheckException(Exception):
